requirements.txt
# Core web framework
flask>=2.2.0
gunicorn>=20.1.0

# HTTP and web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# OpenAI API
openai>=1.0.0

# Basic utilities
python-dotenv>=0.19.0

# Lightweight ML dependencies (only if needed)
numpy>=1.21.0,<1.25.0
scikit-learn>=1.3.0,<1.4.0

# Text processing (lightweight)
nltk>=3.8